refactor all files into packages
