+++ app/src/main/java/com/to_do_list/app/todo/ToDoClickListener.kt
@@ -1,7 +1,7 @@
-package com.to_do_list.app
+package com.to_do_list.app.todo
 
 import com.to_do_list.app.common.entities.ToDo
 
-interface ToDoListClickListener {
+interface ToDoClickListener {
     fun onListClickListener(data: ToDo)
 }
