+++ app/src/main/java/com/to_do_list/app/todolist/EditDeleteToDoListActivity.kt
@@ -1,9 +1,10 @@
-package com.to_do_list.app
+package com.to_do_list.app.todolist
 
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.textfield.TextInputEditText
+import com.to_do_list.app.R
 
 /**
  * @author Nichlas Daniel Boraso(nibor19)
