+++ app/src/main/java/com/to_do_list/app/todolist/ListOnClickListener.kt
@@ -1,4 +1,4 @@
-package com.to_do_list.app
+package com.to_do_list.app.todolist
 
 import com.to_do_list.app.common.entities.ToDoList
