+++ app/src/main/java/com/to_do_list/app/todolist/ToDoListAdapter.kt
@@ -1,4 +1,4 @@
-package com.to_do_list.app
+package com.to_do_list.app.todolist
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
