+++ app/src/main/res/layout/activity_list_of_to_dos.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_grey"
-    tools:context=".ListOfToDos">
+    tools:context=".todo.ToDoOverviewActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar2"
