+++ app/src/main/res/layout/activity_todo_add_update_delete.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".TodoAddUpdateDeleteActivity">
+    tools:context=".todo.TodoAddUpdateDeleteActivity">
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
         android:layout_width="match_parent"
