+++ app/src/main/res/layout/add_list_view.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/backgroundColor"
-    tools:context=".AddToDoLIstActivity">
+    tools:context=".todolist.AddToDoLIstActivity">
 
     <LinearLayout
         android:layout_width="wrap_content"
