+++ app/src/main/res/layout/edit_delete_list_view.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/backgroundColor"
-    tools:context=".EditDeleteToDoListActivity">
+    tools:context=".todolist.EditDeleteToDoListActivity">
 
     <LinearLayout
         android:layout_width="wrap_content"
